fix: treat long as double
This is to be able to fit file sizes (and any other long > 32 bits) in JavaScript numbers. Note that a
whole 64 bit long does not fit into a JavaScript double because some precision is lost. However,
until we find a specific case where this causes an error we will simply treat long as double.