test(proc): fix proc.fork2 test
Waiting on a pid that it's not a child of the current process always lead
to failure so this test was only working just because my machine has
several cores and by the time we ran the expect, the child had already
finished.

However, it randomly failed when on CI because it is a more constrained
environment.